Added new py doc, for graphs
